integer type changes to stifle warnings from the clang static analyzer

+++ dumpbuf.h
@@ -1,1 +1,1 @@
-void DumpBuffer(void*, int, char*);
+void DumpBuffer(void*, long, char*);
